players are now rewarded with block balance when mining certain blocks that are natural

+++ src/main/java/com/ismaiiil/alliance/json/PlayerData.java
@@ -7,8 +7,8 @@ import java.io.Serializable;
 @ToString
 public class PlayerData implements Serializable {
     public int regionsCreated = 0;
-    public int balance = 0;
-    public int usedBalance = 0;
+    public double balance = 0;
+    public double usedBalance = 0;
 
     public PlayerData() {}
 }
